Test S3-based hubs with skip_checks = TRUE
This changeset updates the test suite to test the behavior
of skip_checks = TRUE (which is the default for S3-based hubs).
However, the code as written will not work when there multiple file
types (e.g., csv and parquet), because it performs an Arrow
open_dataset for each file type. That doesn't work when
exclude_invalid_files is FALSE because open_dataset will then
grab every file every time it is run.