use tidyselect 1.2.0 syntax for selection operations
I was running tests for this package when I saw a flood of warnings that all looked like the one at the bottom of this issue.

I then saw https://github.com/hubverse-org/hubData/pull/36#discussion_r1637826949 and found the fix by looking for selecting operations and then whittling down.